Add _pstats module
