+++ pure_utils/__init__.py
@@ -5,4 +5,5 @@ __version__ = "0.4.0"
 from .common import *  # noqa: F401, F403
 from .debug import *  # noqa: F401, F403
 from .dt import *  # noqa: F401, F403
+from .profiler import *  # noqa: F401, F403
 from .strings import *  # noqa: F401, F403
